+++ src/routes/RequestsRoute.js
@@ -64,6 +64,7 @@ import {
   fulfillmentTypeMap,
   DEFAULT_REQUEST_TYPE_VALUE,
   INPUT_REQUEST_SEARCH_SELECTOR,
+  PRINT_DETAILS_COLUMNS,
 } from '../constants';
 import {
   buildUrl,
@@ -98,6 +99,7 @@ import {
   isReorderableRequest,
   getFormattedYears,
   getStatusQuery,
+  getFullNameForCsvRecords,
 } from './utils';
 import SinglePrintButtonForPickSlip from '../components/SinglePrintButtonForPickSlip';
 
@@ -114,8 +116,22 @@ export const getPrintHoldRequestsEnabled = (printHoldRequests) => {
   return printHoldRequestsEnabled;
 };
 
+export const getFilteredColumnHeadersMap = (columnHeaders) => (
+  columnHeaders.filter(column => column.value !== PRINT_DETAILS_COLUMNS.COPIES &&
+    column.value !== PRINT_DETAILS_COLUMNS.PRINTED)
+);
+
 export const extractPickSlipRequestIds = (pickSlipsData) => {
   return pickSlipsData.map(pickSlip => pickSlip['request.requestID']);
+};
+
+export const getLastPrintedDetails = (printDetails, intl) => {
+  const fullName = getFullName(printDetails?.lastPrintRequester);
+  const formattedDate = intl.formatDate(printDetails?.lastPrintedDate);
+  const formattedTime = intl.formatTime(printDetails?.lastPrintedDate);
+  const localizedDateTime = `${formattedDate}${formattedTime ? ', ' : ''}${formattedTime}`;
+
+  return fullName + ' ' + localizedDateTime;
 };
 
 export const urls = {
@@ -202,9 +218,11 @@ export const getListFormatter = (
     setURL,
   },
   {
+    intl,
     selectedRows,
     pickSlipsToCheck,
     pickSlipsData,
+    isViewPrintDetailsEnabled,
     getPrintContentRef,
     pickSlipsPrintTemplate,
     toggleRowSelection,
@@ -227,16 +245,21 @@ export const getListFormatter = (
     </AppIcon>
   ),
   'requester': rq => (rq.requester ? getFullName(rq.requester) : DEFAULT_FORMATTER_VALUE),
-  'singlePrint': rq => (
-    <SinglePrintButtonForPickSlip
-      request={rq}
-      pickSlipsToCheck={pickSlipsToCheck}
-      pickSlipsPrintTemplate={pickSlipsPrintTemplate}
-      onBeforeGetContentForSinglePrintButton={onBeforeGetContentForSinglePrintButton}
-      onBeforePrintForSinglePrintButton={onBeforePrintForSinglePrintButton}
-      pickSlipsData={pickSlipsData}
-      getPrintContentRef={getPrintContentRef}
-    />),
+  'singlePrint': rq => {
+    const singlePrintButtonProps = {
+      request: rq,
+      pickSlipsToCheck,
+      pickSlipsPrintTemplate,
+      onBeforeGetContentForSinglePrintButton,
+      pickSlipsData,
+      getPrintContentRef,
+      ...(isViewPrintDetailsEnabled && {
+        onBeforePrintForSinglePrintButton
+      }),
+    };
+    return (
+      <SinglePrintButtonForPickSlip {...singlePrintButtonProps} />);
+  },
   'requesterBarcode': rq => (rq.requester ? rq.requester.barcode : DEFAULT_FORMATTER_VALUE),
   'requestStatus': rq => (requestStatusesTranslations[rq.status]
     ? <FormattedMessage id={requestStatusesTranslations[rq.status]} />
@@ -246,6 +269,8 @@ export const getListFormatter = (
   'year': rq => getFormattedYears(rq.instance?.publication, DEFAULT_DISPLAYED_YEARS_AMOUNT),
   'callNumber': rq => effectiveCallNumber(rq.item),
   'servicePoint': rq => get(rq, 'pickupServicePoint.name', DEFAULT_FORMATTER_VALUE),
+  'copies': rq => get(rq, PRINT_DETAILS_COLUMNS.COPIES, DEFAULT_FORMATTER_VALUE),
+  'printed': rq => (rq.printDetails ? getLastPrintedDetails(rq.printDetails, intl) : DEFAULT_FORMATTER_VALUE),
 });
 
 export const buildHoldRecords = (records) => {
@@ -262,6 +287,8 @@ export const buildHoldRecords = (records) => {
     return record;
   });
 };
+
+export const viewPrintDetailsPath = 'circulationSettings.records[0].value.enablePrintLog';
 
 class RequestsRoute extends React.Component {
   static contextType = CalloutContext;
@@ -449,6 +476,15 @@ class RequestsRoute extends React.Component {
         query: '(module==SETTINGS and configName==TLR)',
       },
     },
+    circulationSettings: {
+      throwErrors: false,
+      type: 'okapi',
+      records: 'circulationSettings',
+      path: 'circulation/settings',
+      params: {
+        query: '(name=printEventLogFeature)',
+      },
+    },
     savePrintDetails: {
       type: 'okapi',
       POST: {
@@ -503,6 +539,9 @@ class RequestsRoute extends React.Component {
         reset: PropTypes.func.isRequired,
         GET: PropTypes.func.isRequired,
       }).isRequired,
+      circulationSettings: PropTypes.shape({
+        GET: PropTypes.func,
+      }),
       savePrintDetails: PropTypes.shape({
         POST: PropTypes.func,
       }),
@@ -540,6 +579,9 @@ class RequestsRoute extends React.Component {
       printHoldRequests: PropTypes.shape({
         records: PropTypes.arrayOf(PropTypes.object).isRequired,
       }),
+      circulationSettings: PropTypes.shape({
+        records: PropTypes.arrayOf(PropTypes.object),
+      }),
     }).isRequired,
     stripes: PropTypes.shape({
       connect: PropTypes.func.isRequired,
@@ -573,6 +615,8 @@ class RequestsRoute extends React.Component {
       createTitleLevelRequestsByDefault = false,
     } = getTlrSettings(props.resources.configs.records[0]?.value);
 
+    const isViewPrintDetailsEnabled = get(props.resources, viewPrintDetailsPath) === 'true';
+
     this.okapiUrl = props.stripes.okapi.url;
 
     this.httpHeadersOptions = {
@@ -604,6 +648,7 @@ class RequestsRoute extends React.Component {
       selectedRows: {},
       titleLevelRequestsFeatureEnabled,
       createTitleLevelRequestsByDefault,
+      isViewPrintDetailsEnabled,
     };
 
     this.pickSlipsPrintContentRef = React.createRef();
@@ -636,6 +681,17 @@ class RequestsRoute extends React.Component {
     const { configs: prevConfigs } = prevProps.resources;
     const { configs } = this.props.resources;
     const instanceId = parse(this.props.location?.search)?.instanceId;
+    const currPrintDetailsSettings = get(this.props.resources, viewPrintDetailsPath);
+    const prevPrintDetailsSettings = get(prevProps.resources, viewPrintDetailsPath);
+
+    if (currPrintDetailsSettings !== prevPrintDetailsSettings) {
+      const isViewPrintDetailsEnabled = currPrintDetailsSettings === 'true';
+      this.setState({ isViewPrintDetailsEnabled });
+
+      if (!isViewPrintDetailsEnabled) {
+        this.columnHeadersMap = getFilteredColumnHeadersMap(this.columnHeadersMap);
+      }
+    }
 
     if (prevExpired.length > 0 && expired.length === 0) {
       // eslint-disable-next-line react/no-did-update-set-state
@@ -764,7 +820,6 @@ class RequestsRoute extends React.Component {
       queryString = `(requesterId=="${queryTerm}" or requester.barcode="${queryTerm}*" or item.title="${queryTerm}*" or item.barcode=="${queryTerm}*" or itemId=="${queryTerm}")`;
       queryClauses.push(queryString);
     }
-
     if (filterQuery) queryClauses.push(filterQuery);
 
     queryString = queryClauses.join(' and ');
@@ -831,16 +886,21 @@ class RequestsRoute extends React.Component {
         });
       }
       if (record.requester) {
-        const { firstName, middleName, lastName } = record.requester;
-        record.requester.name = `${firstName || ''} ${middleName || ''} ${lastName || ''}`;
+        record.requester.name = getFullNameForCsvRecords(record.requester);
+      }
+      if (record.printDetails) {
+        const fullName = getFullNameForCsvRecords(record.printDetails.lastPrintRequester);
+        const lastPrintedDate = record.printDetails.lastPrintedDate || '';
+        const date = lastPrintedDate ? `, ${lastPrintedDate}` : '';
+
+        record.printDetails.lastPrintedDetails = `${fullName}${date}`;
       }
       if (record.loan) {
         const { dueDate } = record.loan;
         record.loan.dueDate = `${formatDate(dueDate)}, ${formatTime(dueDate)}`;
       }
       if (record.proxy) {
-        const { firstName, middleName, lastName } = record.proxy;
-        record.proxy.name = `${firstName || ''} ${middleName || ''} ${lastName || ''}`;
+        record.proxy.name = getFullNameForCsvRecords(record.proxy);
       }
       if (record.deliveryAddress) {
         const { addressLine1, city, region, postalCode, countryId } = record.deliveryAddress;
@@ -1190,7 +1250,8 @@ class RequestsRoute extends React.Component {
   };
 
   savePrintEventDetails = (requestIds) => {
-    const printTimeStamp = moment.tz('UTC').format();
+    const currDateTime = new Date();
+    const printTimeStamp = currDateTime.toISOString();
     const { id: loggedInUserId, username: loggedInUsername } = this.props.stripes.user.user;
 
     this.props.mutator.savePrintDetails.POST({
@@ -1211,23 +1272,12 @@ class RequestsRoute extends React.Component {
     })
   );
 
-  onBeforePrintForPrintButton = (slipType, slipsData) => {
-    if (slipType === SLIPS_TYPE.PICK_SLIP) { // TODO: add enable print log settings check
-      const requestIds = extractPickSlipRequestIds(slipsData);
-      this.savePrintEventDetails(requestIds);
-    }
-  };
-
   onBeforeGetContentForSinglePrintButton = () => (
     new Promise(resolve => {
       this.context.sendCallout({ message: <FormattedMessage id="ui-requests.printInProgress" /> });
       setTimeout(() => resolve(), 1000);
     })
   );
-
-  onBeforePrintForSinglePrintButton = (requestId) => {
-    this.savePrintEventDetails([requestId]); // TODO: add enable print log settings check
-  }
 
   printContentRefs = {};
 
@@ -1260,30 +1310,6 @@ class RequestsRoute extends React.Component {
   }
 
   render() {
-    const columnLabels = {
-      select: <Checkbox
-        data-testid="selectRequestCheckbox"
-        checked={this.getIsAllRowsSelected()}
-        aria-label={<FormattedMessage id="ui-requests.instances.rows.select" />}
-        onChange={this.toggleAllRows}
-      />,
-      requestDate: <FormattedMessage id="ui-requests.requests.requestDate" />,
-      title: <FormattedMessage id="ui-requests.requests.title" />,
-      year: <FormattedMessage id="ui-requests.requests.year" />,
-      itemBarcode: <FormattedMessage id="ui-requests.requests.itemBarcode" />,
-      callNumber: <FormattedMessage id="ui-requests.requests.callNumber" />,
-      type: <FormattedMessage id="ui-requests.requests.type" />,
-      requestStatus: <FormattedMessage id="ui-requests.requests.status" />,
-      position: <FormattedMessage id="ui-requests.requests.queuePosition" />,
-      servicePoint: <FormattedMessage id="ui-requests.requests.servicePoint" />,
-      requester: <FormattedMessage id="ui-requests.requests.requester" />,
-      requesterBarcode: <FormattedMessage id="ui-requests.requests.requesterBarcode" />,
-      singlePrint: <FormattedMessage id="ui-requests.requests.singlePrint" />,
-      proxy: <FormattedMessage id="ui-requests.requests.proxy" />,
-      copies: <FormattedMessage id="ui-requests.requests.copies" />,
-      printed: <FormattedMessage id="ui-requests.requests.printed" />,
-    };
-
     const {
       resources,
       mutator,
@@ -1308,6 +1334,7 @@ class RequestsRoute extends React.Component {
       selectedRows,
       holdsShelfReportPending,
       createTitleLevelRequestsByDefault,
+      isViewPrintDetailsEnabled,
     } = this.state;
     const isPrintHoldRequestsEnabled = getPrintHoldRequestsEnabled(resources.printHoldRequests);
     const { name: servicePointName } = this.getCurrentServicePointInfo();
@@ -1323,6 +1350,32 @@ class RequestsRoute extends React.Component {
       requestType: DEFAULT_REQUEST_TYPE_VALUE,
       fulfillmentPreference: fulfillmentTypeMap.HOLD_SHELF,
       createTitleLevelRequest: createTitleLevelRequestsByDefault,
+    };
+
+    const columnLabels = {
+      select: <Checkbox
+        data-testid="selectRequestCheckbox"
+        checked={this.getIsAllRowsSelected()}
+        aria-label={<FormattedMessage id="ui-requests.instances.rows.select" />}
+        onChange={this.toggleAllRows}
+      />,
+      requestDate: <FormattedMessage id="ui-requests.requests.requestDate" />,
+      title: <FormattedMessage id="ui-requests.requests.title" />,
+      year: <FormattedMessage id="ui-requests.requests.year" />,
+      itemBarcode: <FormattedMessage id="ui-requests.requests.itemBarcode" />,
+      callNumber: <FormattedMessage id="ui-requests.requests.callNumber" />,
+      type: <FormattedMessage id="ui-requests.requests.type" />,
+      requestStatus: <FormattedMessage id="ui-requests.requests.status" />,
+      position: <FormattedMessage id="ui-requests.requests.queuePosition" />,
+      servicePoint: <FormattedMessage id="ui-requests.requests.servicePoint" />,
+      requester: <FormattedMessage id="ui-requests.requests.requester" />,
+      requesterBarcode: <FormattedMessage id="ui-requests.requests.requesterBarcode" />,
+      singlePrint: <FormattedMessage id="ui-requests.requests.singlePrint" />,
+      proxy: <FormattedMessage id="ui-requests.requests.proxy" />,
+      ...(isViewPrintDetailsEnabled && {
+        copies: <FormattedMessage id="ui-requests.requests.copies" />,
+        printed: <FormattedMessage id="ui-requests.requests.printed" />,
+      }),
     };
 
     const isPickSlipsArePending = resources?.pickSlips?.isPending;
@@ -1342,14 +1395,16 @@ class RequestsRoute extends React.Component {
         setURL: this.setURL,
       },
       {
+        intl,
         selectedRows,
         pickSlipsToCheck: pickSlips,
         pickSlipsData,
+        isViewPrintDetailsEnabled,
         getPrintContentRef: this.getPrintContentRef,
         pickSlipsPrintTemplate,
         toggleRowSelection: this.toggleRowSelection,
         onBeforeGetContentForSinglePrintButton: this.onBeforeGetContentForSinglePrintButton,
-        onBeforePrintForSinglePrintButton: this.onBeforePrintForSinglePrintButton,
+        onBeforePrintForSinglePrintButton: this.savePrintEventDetails,
       }
     );
 
@@ -1413,7 +1468,12 @@ class RequestsRoute extends React.Component {
                   template={pickSlipsPrintTemplate}
                   contentRef={this.pickSlipsPrintContentRef}
                   onBeforeGetContent={() => this.onBeforeGetContentForPrintButton(onToggle)}
-                  onBeforePrint={() => this.onBeforePrintForPrintButton(SLIPS_TYPE.PICK_SLIP, pickSlipsData)}
+                  onBeforePrint={() => {
+                    if (isViewPrintDetailsEnabled) {
+                      const requestIds = extractPickSlipRequestIds(pickSlipsData);
+                      this.savePrintEventDetails(requestIds);
+                    }
+                  }}
                 >
                   <FormattedMessage
                     id="ui-requests.printPickSlips"
@@ -1438,9 +1498,11 @@ class RequestsRoute extends React.Component {
                     }
                   onBeforePrint={
                     () => {
-                      const selectedPickSlips = getSelectedSlipDataMulti(pickSlipsData, selectedRows);
-                      const pickSlipsRequestIds = extractPickSlipRequestIds(selectedPickSlips);
-                      this.savePrintEventDetails(pickSlipsRequestIds); // TODO: add enable print log settings check
+                      if (isViewPrintDetailsEnabled) {
+                        const selectedPickSlips = getSelectedSlipDataMulti(pickSlipsData, selectedRows);
+                        const selectedRequestIds = extractPickSlipRequestIds(selectedPickSlips);
+                        this.savePrintEventDetails(selectedRequestIds);
+                      }
                     }
                   }
                 >
@@ -1466,7 +1528,6 @@ class RequestsRoute extends React.Component {
                     template={searchSlipsPrintTemplate}
                     contentRef={this.searchSlipsPrintContentRef}
                     onBeforeGetContent={() => this.onBeforeGetContentForPrintButton(onToggle)}
-                    onBeforePrint={() => this.onBeforePrintForPrintButton(SLIPS_TYPE.SEARCH_SLIP_HOLD_REQUESTS)}
                   >
                     <FormattedMessage
                       id="ui-requests.printSearchSlips"