+++ src/routes/RequestsRoute.test.js
@@ -2,6 +2,7 @@ import React from 'react';
 import {
   stringify,
 } from 'query-string';
+import { createIntl, createIntlCache } from 'react-intl';
 
 import {
   render,
@@ -33,6 +34,8 @@ import RequestsRoute, {
   buildHoldRecords,
   getListFormatter,
   getPrintHoldRequestsEnabled,
+  getLastPrintedDetails,
+  getFilteredColumnHeadersMap,
   urls,
   DEFAULT_FORMATTER_VALUE,
 } from './RequestsRoute';
@@ -61,6 +64,7 @@ import {
   MAX_RECORDS,
   REQUEST_OPERATIONS,
   INPUT_REQUEST_SEARCH_SELECTOR,
+  PRINT_DETAILS_COLUMNS,
 } from '../constants';
 import { historyData } from '../../test/jest/fixtures/historyData';
 
@@ -80,6 +84,15 @@ const testIds = {
   rowCheckbox: 'rowCheckbox',
   selectRequestCheckbox: 'selectRequestCheckbox',
 };
+
+const intlCache = createIntlCache();
+const intl = createIntl(
+  {
+    locale: 'en-US',
+    messages: {},
+  },
+  intlCache
+);
 
 jest.spyOn(React, 'createRef').mockReturnValue(createRefMock);
 jest.spyOn(document, 'getElementById').mockReturnValue(createDocumentRefMock);
@@ -158,7 +171,7 @@ jest.mock('../components/SinglePrintButtonForPickSlip', () => jest.fn(({
 }) => {
   const handleClick = () => {
     onBeforeGetContentForSinglePrintButton();
-    onBeforePrintForSinglePrintButton();
+    onBeforePrintForSinglePrintButton(['reqId']);
   };
   return (
     <button
@@ -217,6 +230,11 @@ const mockedRequest = {
     barcode: 'requesterBarcode',
   },
   id: 'requestId',
+};
+const printDetailsMockData = {
+  count: 11,
+  lastPrintedDate: '2024-08-03T13:33:31.868Z',
+  lastPrintRequester: { firstName: 'firstName', middleName: 'middleName', lastName: 'lastName' },
 };
 
 SearchAndSort.mockImplementation(jest.fn(({
@@ -339,6 +357,7 @@ describe('RequestsRoute', () => {
         postalCode: '12345',
         countryId: 'US',
       },
+      printDetails: printDetailsMockData,
     }
   ];
   const defaultProps = {
@@ -352,6 +371,9 @@ describe('RequestsRoute', () => {
       url: '{{ env.FOLIO_MD_REGISTRY }}/_/proxy/modules',
     },
     mutator: {
+      circulationSettings: {
+        GET: jest.fn(),
+      },
       savePrintDetails: {
         POST: jest.fn().mockResolvedValue(),
       },
@@ -446,6 +468,13 @@ describe('RequestsRoute', () => {
         records: [{
           value: '{"printHoldRequestsEnabled": true}',
         }],
+      },
+      circulationSettings: {
+        records: [{
+          value: {
+            enablePrintLog: 'true',
+          }
+        }]
       },
       currentServicePoint: {},
       patronBlocks: {
@@ -575,11 +604,37 @@ describe('RequestsRoute', () => {
       expect(printContent).toBeInTheDocument();
     });
 
-    it('should trigger "exportCsv"', async () => {
+    it('should trigger "exportCsv" when "Enable view print details" settings is enabled', async () => {
       await userEvent.click(screen.getByRole('button', { name: 'ui-requests.exportSearchResultsToCsv' }));
 
       await waitFor(() => {
-        expect(exportCsv).toBeCalled();
+        expect(exportCsv).toHaveBeenCalled();
+      });
+    });
+
+    it('should trigger "exportCsv" when "Enable view print details" settings is disabled', async () => {
+      const props = {
+        ...defaultProps,
+        resources: {
+          ...defaultProps.resources,
+          circulationSettings: {
+            ...defaultProps.resources.circulationSettings,
+            records: defaultProps.resources.circulationSettings.records.map(record => ({
+              ...record,
+              value: {
+                ...record.value,
+                enablePrintLog: 'false'
+              }
+            }))
+          },
+        }
+      };
+
+      renderComponent(props);
+      await userEvent.click(screen.queryAllByRole('button', { name: 'ui-requests.exportSearchResultsToCsv' })[1]);
+
+      await waitFor(() => {
+        expect(exportCsv).toHaveBeenCalled();
       });
     });
 
@@ -632,6 +687,12 @@ describe('RequestsRoute', () => {
       await userEvent.click(screen.getAllByRole('button', { name: 'PrintButton' })[0]);
 
       expect(printPickSlipsLabel).toBeInTheDocument();
+    });
+
+    it('should trigger "mutator.savePrintDetails.POST"', async () => {
+      await userEvent.click(screen.getAllByRole('button', { name: 'PrintButton' })[0]);
+
+      expect(defaultProps.mutator.savePrintDetails.POST).toHaveBeenCalled();
     });
 
     it('should render print search slips label', async () => {
@@ -1002,9 +1063,11 @@ describe('RequestsRoute', () => {
         setURL: setURLMock,
       },
       {
+        intl,
         selectedRows: '',
         pickSlipsToCheck: '',
         pickSlipsData: '',
+        isViewPrintDetailsEnabled: true,
         getPrintContentRef: getPrintContentRefMock,
         isPrintableMock,
         pickSlipsPrintTemplate: '',
@@ -1037,6 +1100,7 @@ describe('RequestsRoute', () => {
       pickupServicePoint: {
         name: 'name',
       },
+      printDetails: printDetailsMockData,
     };
     const requestWithoutData = {};
 
@@ -1044,6 +1108,7 @@ describe('RequestsRoute', () => {
       it('should render CheckboxColumn', () => {
         const selectedRequest = {};
         const options = {
+          intl,
           selectedRows: [],
           pickSlipsToCheck: [],
           pickSlipsData: {},
@@ -1066,6 +1131,7 @@ describe('RequestsRoute', () => {
       it('should render SinglePrintButtonForPickSlip', () => {
         const selectedRequest = {};
         const options = {
+          intl,
           selectedRows: [],
           pickSlipsToCheck: [],
           pickSlipsData: {},
@@ -1235,6 +1301,29 @@ describe('RequestsRoute', () => {
         expect(listFormatter.servicePoint(requestWithData)).toBe(requestWithData.pickupServicePoint.name);
       });
     });
+
+    describe('when formatting copies column', () => {
+      it('should return copies for copies column', () => {
+        expect(listFormatter.copies(requestWithData)).toBe(requestWithData.printDetails.count);
+      });
+    });
+
+    describe('when formatting printed column', () => {
+      it('should return last printed details for printed column', () => {
+        getFullName.mockReturnValueOnce('lastName, firstName middleName');
+
+        const expectedFormattedDate = intl.formatDate(requestWithData.printDetails.lastPrintedDate);
+        const expectedFormattedTime = intl.formatTime(requestWithData.printDetails.lastPrintedDate);
+        const expectedOutput =
+        `lastName, firstName middleName ${expectedFormattedDate}${expectedFormattedTime ? ', ' : ''}${expectedFormattedTime}`;
+
+        expect(listFormatter.printed(requestWithData)).toBe(expectedOutput);
+      });
+
+      it('should return empty string', () => {
+        expect(listFormatter.printed(requestWithoutData)).toBe(DEFAULT_FORMATTER_VALUE);
+      });
+    });
   });
 
   describe('getPrintHoldRequestsEnabled', () => {
@@ -1258,6 +1347,49 @@ describe('RequestsRoute', () => {
       expect(getPrintHoldRequestsEnabled({
         records: [],
       })).toBeFalsy();
+    });
+  });
+
+  describe('getLastPrintedDetails', () => {
+    const lastPrintDetails = {
+      lastPrintRequester: { firstName: 'firstName', middleName: 'middleName', lastName: 'lastName' },
+      lastPrintedDate: '2024-08-03T13:33:31.868Z',
+    };
+    beforeEach(() => {
+      getFullName.mockClear().mockReturnValue('lastName, firstName middleName');
+    });
+
+    it('should render getLastPrintedDetails() correctly', () => {
+      expect(getLastPrintedDetails(lastPrintDetails, intl)).toBeTruthy();
+    });
+
+    it('should return the formatted full name and date/time correctly', () => {
+      const printedDetails = getLastPrintedDetails(lastPrintDetails, intl);
+      const expectedFormattedDate = intl.formatDate(lastPrintDetails.lastPrintedDate);
+      const expectedFormattedTime = intl.formatTime(lastPrintDetails.lastPrintedDate);
+      const expectedOutput =
+        `lastName, firstName middleName ${expectedFormattedDate}${expectedFormattedTime ? ', ' : ''}${expectedFormattedTime}`;
+
+      expect(printedDetails).toBe(expectedOutput);
+    });
+  });
+
+  describe('getFilteredColumnHeadersMap', () => {
+    const columnHeaders = [
+      { value: 'patronComments' },
+      { value: PRINT_DETAILS_COLUMNS.COPIES },
+      { value: PRINT_DETAILS_COLUMNS.PRINTED },
+    ];
+
+    it('should render getFilteredColumnHeadersMap() correctly', () => {
+      expect(getFilteredColumnHeadersMap(columnHeaders)).toBeTruthy();
+    });
+
+    it('should return properly filtered column headers', () => {
+      const filteredColumnHeaders = getFilteredColumnHeadersMap(columnHeaders);
+      const expectedColumnHeaders = [{ value: 'patronComments' }];
+
+      expect(filteredColumnHeaders).toEqual(expectedColumnHeaders);
     });
   });